restored shadow manager, trying to resolve circular dependencies
