+++ src/shaders/debug_shadow_shader.cpp
@@ -26,7 +26,7 @@ void DebugShadowShader::post_draw() {}
 void DebugShadowShader::draw(Mesh *mesh, SceneInfo &scene_info, glm::mat4 model_to_world) {
     // Set texture to be shadow map texture.
     glActiveTexture(GL_TEXTURE0);
-    glBindTexture(GL_TEXTURE_2D, Shader::shadow.shadow_map_tex);
+    glBindTexture(GL_TEXTURE_2D, ShaderManager::shadow.shadow_map_tex);
     // Draw quad.
     glBindVertexArray(VAO);
     glDrawArrays(GL_TRIANGLE_STRIP, 0, 4);
